Update end of beacon malfunction text
